refactor(cpp): split completion function / state to new struct
