+++ cpp/rn-llama.cpp
@@ -1,6 +1,7 @@
 #include "rn-llama.h"
 #include "rn-tts.h"
 #include "rn-mtmd.hpp"
+#include "rn-completion.h"
 
 // Include multimodal support
 #include "tools/mtmd/mtmd.h"
@@ -36,9 +37,6 @@ lm_ggml_type kv_cache_type_from_str(const std::string & s) {
     return LM_GGML_TYPE_F16;
 }
 
-static void llama_batch_clear(llama_batch *batch) {
-    batch->n_tokens = 0;
-}
 
 void log(const char *level, const char *function, int line,
                        const char *format, ...)
@@ -125,45 +123,13 @@ std::string tokens_to_str(llama_context *ctx, const std::vector<llama_token>::co
 
 
 llama_rn_context::~llama_rn_context() {
-    if (ctx_sampling != nullptr) {
-        common_sampler_free(ctx_sampling);
+    if (completion != nullptr) {
+        delete completion;
+        completion = nullptr;
     }
 
     releaseMultimodal();
     releaseVocoder();
-}
-
-void llama_rn_context::rewind() {
-    is_interrupted = false;
-    params.antiprompt.clear();
-    params.sampling.grammar.clear();
-    num_prompt_tokens = 0;
-    num_tokens_predicted = 0;
-    generated_text = "";
-    generated_text.reserve(params.n_ctx);
-    truncated = false;
-    context_full = false;
-    stopped_eos = false;
-    stopped_word = false;
-    stopped_limit = false;
-    stopping_word = "";
-    incomplete = false;
-    n_remain = 0;
-    n_past = 0;
-    params.sampling.n_prev = n_ctx;
-    if (isVocoderEnabled()) {
-        tts_wrapper->audio_tokens.clear();
-        tts_wrapper->next_token_uses_guide_token = true;
-        tts_wrapper->guide_tokens.clear();
-    }
-}
-
-bool llama_rn_context::initSampling() {
-    if (ctx_sampling != nullptr) {
-        common_sampler_free(ctx_sampling);
-    }
-    ctx_sampling = common_sampler_init(model, params.sampling);
-    return ctx_sampling != nullptr;
 }
 
 bool llama_rn_context::loadModel(common_params &params_)
@@ -180,6 +146,12 @@ bool llama_rn_context::loadModel(common_params &params_)
     templates = common_chat_templates_init(model, params.chat_template);
     n_ctx = llama_n_ctx(ctx);
 
+    // Initialize completion context
+    if (completion != nullptr) {
+        delete completion;
+    }
+    completion = new llama_rn_context_completion(this);
+
     // Initialize context shift flag
     LOG_INFO("ctx_shift: %s", params.ctx_shift ? "enabled" : "disabled");
 
@@ -188,6 +160,7 @@ bool llama_rn_context::loadModel(common_params &params_)
 
     return true;
 }
+
 
 bool llama_rn_context::validateModelChatTemplate(bool use_jinja, const char *name) const {
     const char * tmpl = llama_model_chat_template(model, name);
@@ -266,588 +239,6 @@ std::string llama_rn_context::getFormattedChat(
     }
 }
 
-void llama_rn_context::truncatePrompt(std::vector<llama_token> &prompt_tokens) {
-    const int n_left = n_ctx - params.n_keep;
-    const int n_block_size = n_left / 2;
-    const int erased_blocks = (prompt_tokens.size() - params.n_keep - n_block_size) / n_block_size;
-
-    // Keep n_keep tokens at start of prompt (at most n_ctx - 4)
-    std::vector<llama_token> new_tokens(prompt_tokens.begin(), prompt_tokens.begin() + params.n_keep);
-
-    new_tokens.insert(new_tokens.end(), prompt_tokens.begin() + params.n_keep + erased_blocks * n_block_size, prompt_tokens.end());
-
-    LOG_INFO("input truncated, n_ctx: %d, n_keep: %d, n_left: %d, old_size: %d, new_size: %d",
-        n_ctx,
-        params.n_keep,
-        n_left,
-        prompt_tokens.size(),
-        new_tokens.size()
-    );
-
-    truncated = true;
-    prompt_tokens = new_tokens;
-}
-
-void llama_rn_context::loadPrompt(const std::vector<std::string> &media_paths) {
-    bool has_media = !media_paths.empty();
-
-    if (!has_media) {
-        std::vector<llama_token> text_tokens;
-        // Text-only path
-        text_tokens = ::common_tokenize(ctx, params.prompt, true, true);
-        num_prompt_tokens = text_tokens.size();
-
-        // LOG tokens
-        std::stringstream ss;
-        ss << "\n" << __func__ << ": prompt_tokens = ";
-        for (auto& token : text_tokens) {
-            ss << token << " ";
-        }
-        LOG_INFO("%s\n", ss.str().c_str());
-
-        if (params.n_keep < 0) {
-            params.n_keep = (int)num_prompt_tokens;
-        }
-        params.n_keep = std::min(n_ctx - 4, params.n_keep);
-
-        // Handle truncation if needed
-        if (num_prompt_tokens >= (size_t)n_ctx) {
-            if (!params.ctx_shift) {
-                context_full = true;
-                return;
-            }
-            truncatePrompt(text_tokens);
-            num_prompt_tokens = text_tokens.size();
-            LM_GGML_ASSERT(num_prompt_tokens < (size_t)n_ctx);
-        }
-
-        // Update sampling context
-        for (auto & token : text_tokens) {
-            common_sampler_accept(ctx_sampling, token, false);
-        }
-
-        // compare the evaluated prompt with the new prompt
-        n_past = common_part(embd, text_tokens);
-
-        embd = text_tokens;
-        if (n_past == num_prompt_tokens) {
-            // we have to evaluate at least 1 token to generate logits.
-            n_past--;
-        }
-
-        // Manage KV cache
-        auto * kv = llama_get_memory(ctx);
-        llama_memory_seq_rm(kv, 0, n_past, -1);
-
-        LOG_VERBOSE("prompt ingested, n_past: %d, cached: %s, to_eval: %s",
-            n_past,
-            tokens_to_str(ctx, embd.cbegin(), embd.cbegin() + n_past).c_str(),
-            tokens_to_str(ctx, embd.cbegin() + n_past, embd.cend()).c_str()
-        );
-    } else {
-        // Multimodal path - process all media paths
-        processMedia(params.prompt, media_paths);
-        num_prompt_tokens = embd.size();
-    }
-
-    has_next_token = true;
-
-    LOG_INFO("[DEBUG] Input processed: n_past=%d, embd.size=%zu, num_prompt_tokens=%zu, has_media=%d",
-             n_past, embd.size(), num_prompt_tokens, has_media ? 1 : 0);
-}
-
-
-void llama_rn_context::beginCompletion() {
-    beginCompletion(COMMON_CHAT_FORMAT_CONTENT_ONLY, COMMON_REASONING_FORMAT_NONE, false);
-}
-
-void llama_rn_context::beginCompletion(int chat_format, common_reasoning_format reasoning_format, bool thinking_forced_open) {
-    // number of tokens to keep when resetting context
-    n_remain = params.n_predict;
-    llama_perf_context_reset(ctx);
-    is_predicting = true;
-
-    current_chat_format = chat_format;
-    current_reasoning_format = reasoning_format;
-    current_thinking_forced_open = thinking_forced_open;
-}
-
-void llama_rn_context::endCompletion() {
-    is_predicting = false;
-}
-
-completion_token_output llama_rn_context::nextToken()
-{
-    completion_token_output result;
-    result.tok = -1;
-
-    if (embd.size() >= (size_t)params.n_ctx)
-    {
-        if (!params.ctx_shift) {
-            // If context shifting is disabled, stop generation
-            LOG_WARNING("context full, n_ctx: %d, tokens: %d", params.n_ctx, embd.size());
-            has_next_token = false;
-            context_full = true;
-            return result;
-        }
-
-        // Shift context
-
-        const int n_left    = n_past - params.n_keep - 1;
-        const int n_discard = n_left/2;
-
-        auto * kv = llama_get_memory(ctx);
-        llama_memory_seq_rm (kv, 0, params.n_keep + 1            , params.n_keep + n_discard + 1);
-        llama_memory_seq_add(kv, 0, params.n_keep + 1 + n_discard, n_past, -n_discard);
-
-        for (size_t i = params.n_keep + 1 + n_discard; i < embd.size(); i++)
-        {
-            embd[i - n_discard] = embd[i];
-        }
-        embd.resize(embd.size() - n_discard);
-
-        n_past -= n_discard;
-        truncated = true;
-
-        LOG_VERBOSE("context shifted, new n_past: %d, new size: %d", n_past, embd.size());
-    }
-
-    bool tg = true;
-    while (n_past < embd.size())
-    {
-        int n_eval = (int)embd.size() - n_past;
-        tg = n_eval == 1;
-        if (n_eval > params.n_batch)
-        {
-            n_eval = params.n_batch;
-        }
-        if (llama_decode(ctx, llama_batch_get_one(&embd[n_past], n_eval)))
-        {
-            LOG_ERROR("failed to eval, n_eval: %d, n_past: %d, n_threads: %d, embd: %s",
-                n_eval,
-                n_past,
-                params.cpuparams.n_threads,
-                tokens_to_str(ctx, embd.cbegin() + n_past, embd.cend()).c_str()
-            );
-            has_next_token = false;
-            return result;
-        }
-        n_past += n_eval;
-
-        if(is_interrupted) {
-            LOG_INFO("Decoding Interrupted");
-            embd.resize(n_past);
-            has_next_token = false;
-            return result;
-        }
-    }
-
-    const llama_vocab* vocab = llama_model_get_vocab(model);
-
-    if (params.n_predict == 0)
-    {
-        has_next_token = false;
-        result.tok = llama_vocab_eos(vocab);
-        return result;
-    }
-
-    {
-        // out of user input, sample next token
-        std::vector<llama_token_data> candidates;
-        candidates.reserve(llama_vocab_n_tokens(vocab));
-
-        llama_token new_token_id = common_sampler_sample(ctx_sampling, ctx, -1);
-
-        if (llama_vocab_is_eog(vocab, new_token_id)) {
-            has_next_token = false;
-            stopped_eos = true;
-            LOG_VERBOSE("EOS: %s", common_token_to_piece(ctx, new_token_id).c_str());
-            return result;
-        }
-
-        if (tts_wrapper != nullptr && tts_wrapper->next_token_uses_guide_token && !tts_wrapper->guide_tokens.empty() && !llama_vocab_is_control(vocab, new_token_id)) {
-            new_token_id = tts_wrapper->guide_tokens[0];
-            tts_wrapper->guide_tokens.erase(tts_wrapper->guide_tokens.begin());
-        }
-        if (tts_wrapper != nullptr) {
-            tts_wrapper->next_token_uses_guide_token = (new_token_id == 198);
-        }
-        result.tok = new_token_id;
-
-        llama_token_data_array cur_p = *common_sampler_get_candidates(ctx_sampling);
-
-        const int32_t n_probs = params.sampling.n_probs;
-
-        // deprecated
-        /*if (params.sampling.temp <= 0 && n_probs > 0)
-        {
-            // For llama_sample_token_greedy we need to sort candidates
-            llama_sampler_init_softmax();
-
-        }*/
-
-
-        for (size_t i = 0; i < std::min(cur_p.size, (size_t)n_probs); ++i)
-        {
-            result.probs.push_back({cur_p.data[i].id, cur_p.data[i].p});
-        }
-
-        common_sampler_accept(ctx_sampling, result.tok, true);
-        if (tg) {
-            num_tokens_predicted++;
-        }
-    }
-
-    // add it to the context
-    embd.push_back(result.tok);
-    // decrement remaining sampling budget
-    --n_remain;
-
-    has_next_token = params.n_predict == -1 || n_remain != 0;
-    return result;
-}
-
-size_t llama_rn_context::findStoppingStrings(const std::string &text, const size_t last_token_size,
-                            const stop_type type)
-{
-    size_t stop_pos = std::string::npos;
-    for (const std::string &word : params.antiprompt)
-    {
-        size_t pos;
-        if (type == STOP_FULL)
-        {
-            const size_t tmp = word.size() + last_token_size;
-            const size_t from_pos = text.size() > tmp ? text.size() - tmp : 0;
-            pos = text.find(word, from_pos);
-        }
-        else
-        {
-            pos = find_partial_stop_string(word, text);
-        }
-        if (pos != std::string::npos &&
-            (stop_pos == std::string::npos || pos < stop_pos))
-        {
-            if (type == STOP_FULL)
-            {
-                stopping_word = word;
-                stopped_word = true;
-                has_next_token = false;
-            }
-            stop_pos = pos;
-        }
-    }
-    return stop_pos;
-}
-
-completion_token_output llama_rn_context::doCompletion()
-{
-    completion_token_output token_with_probs = nextToken();
-
-    const std::string token_text = token_with_probs.tok == -1 ? "" : common_token_to_piece(ctx, token_with_probs.tok);
-    generated_text += token_text;
-
-    if (isVocoderEnabled()) {
-        tts_type type = tts_wrapper->getTTSType(this);
-        if (tts_wrapper->type == UNKNOWN) {
-            tts_wrapper->type = type;
-        }
-        if ((type == OUTETTS_V0_2 || type == OUTETTS_V0_3) && (token_with_probs.tok >= 151672 && token_with_probs.tok <= 155772)) {
-            tts_wrapper->audio_tokens.push_back(token_with_probs.tok);
-        }
-    }
-
-    if (params.sampling.n_probs > 0)
-    {
-        generated_token_probs.push_back(token_with_probs);
-    }
-
-    // check if there is incomplete UTF-8 character at the end
-    for (unsigned i = 1; i < 5 && i <= generated_text.size(); ++i) {
-        unsigned char c = generated_text[generated_text.size() - i];
-        if ((c & 0xC0) == 0x80) {
-            // continuation byte: 10xxxxxx
-            continue;
-        }
-        if ((c & 0xE0) == 0xC0) {
-            // 2-byte character: 110xxxxx ...
-            incomplete = i < 2;
-        } else if ((c & 0xF0) == 0xE0) {
-            // 3-byte character: 1110xxxx ...
-            incomplete = i < 3;
-        } else if ((c & 0xF8) == 0xF0) {
-            // 4-byte character: 11110xxx ...
-            incomplete = i < 4;
-        }
-        // else 1-byte character or invalid byte
-        break;
-    }
-
-    if (incomplete && !has_next_token)
-    {
-        has_next_token = true;
-        n_remain++;
-    }
-
-    if (!has_next_token && n_remain == 0)
-    {
-        stopped_limit = true;
-    }
-
-    LOG_VERBOSE("next token, token: %s, token_text: %s, has_next_token: %d, n_remain: %d, num_tokens_predicted: %d, stopped_eos: %d, stopped_word: %d, stopped_limit: %d, stopping_word: %s",
-        common_token_to_piece(ctx, token_with_probs.tok),
-        tokens_to_output_formatted_string(ctx, token_with_probs.tok).c_str(),
-        has_next_token,
-        n_remain,
-        num_tokens_predicted,
-        stopped_eos,
-        stopped_word,
-        stopped_limit,
-        stopping_word.c_str()
-    );
-    return token_with_probs;
-}
-
-completion_partial_output llama_rn_context::getPartialOutput(const std::string &token_text) {
-    common_chat_syntax syntax;
-    syntax.format = static_cast<common_chat_format>(current_chat_format);
-    syntax.reasoning_format = current_reasoning_format;
-    syntax.thinking_forced_open = current_thinking_forced_open;
-    syntax.parse_tool_calls = true;
-
-    common_chat_msg parsed_msg = common_chat_parse(generated_text, true, syntax);
-
-    completion_partial_output result;
-
-    result.content = parsed_msg.content;
-    result.reasoning_content = parsed_msg.reasoning_content;
-    result.accumulated_text = generated_text;
-    result.tool_calls = parsed_msg.tool_calls;
-
-    return result;
-}
-
-std::vector<float> llama_rn_context::getEmbedding(common_params &embd_params)
-{
-    static const int n_embd = llama_model_n_embd(llama_get_model(ctx));
-    if (!embd_params.embedding)
-    {
-        LOG_WARNING("embedding disabled, embedding: %s", embd_params.embedding);
-        return std::vector<float>(n_embd, 0.0f);
-    }
-    float *data;
-
-    const enum llama_pooling_type pooling_type = llama_pooling_type(ctx);
-    printf("pooling_type: %d\n", pooling_type);
-    if (pooling_type == LLAMA_POOLING_TYPE_NONE) {
-        data = llama_get_embeddings(ctx);
-    } else {
-        data = llama_get_embeddings_seq(ctx, 0);
-    }
-
-    if (!data) {
-        return std::vector<float>(n_embd, 0.0f);
-    }
-    std::vector<float> embedding(data, data + n_embd), out(data, data + n_embd);
-    common_embd_normalize(embedding.data(), out.data(), n_embd, embd_params.embd_normalize);
-    return out;
-}
-
-// Helper function to format rerank task: [BOS]query[EOS][SEP]doc[EOS]
-static std::vector<llama_token> format_rerank(const llama_vocab * vocab, const std::vector<llama_token> & query, const std::vector<llama_token> & doc) {
-    std::vector<llama_token> result;
-
-    // Get EOS token - use SEP token as fallback if EOS is not available
-    llama_token eos_token = llama_vocab_eos(vocab);
-    if (eos_token == LLAMA_TOKEN_NULL) {
-        eos_token = llama_vocab_sep(vocab);
-    }
-
-    result.reserve(doc.size() + query.size() + 4);
-    if (llama_vocab_get_add_bos(vocab)) {
-        result.push_back(llama_vocab_bos(vocab));
-    }
-    result.insert(result.end(), query.begin(), query.end());
-    if (llama_vocab_get_add_eos(vocab)) {
-        result.push_back(eos_token);
-    }
-    if (llama_vocab_get_add_sep(vocab)) {
-        result.push_back(llama_vocab_sep(vocab));
-    }
-    result.insert(result.end(), doc.begin(), doc.end());
-    if (llama_vocab_get_add_eos(vocab)) {
-        result.push_back(eos_token);
-    }
-
-    return result;
-}
-
-std::vector<float> llama_rn_context::rerank(const std::string &query, const std::vector<std::string> &documents)
-{
-    std::vector<float> scores;
-
-    // Check if this model supports reranking (requires rank pooling type)
-    const enum llama_pooling_type pooling_type = llama_pooling_type(ctx);
-    if (pooling_type != LLAMA_POOLING_TYPE_RANK) {
-        throw std::runtime_error("reranking not supported, pooling_type: " + std::to_string(pooling_type));
-    }
-
-    if (!params.embedding) {
-        throw std::runtime_error("embedding disabled but required for reranking");
-    }
-
-    const llama_vocab * vocab = llama_model_get_vocab(model);
-    std::vector<llama_token> query_tokens = common_tokenize(vocab, query, false, true);
-
-    scores.reserve(documents.size());
-
-    for (size_t i = 0; i < documents.size(); ++i) {
-        rewind();
-        embd = {};
-
-        const std::string & document = documents[i];
-
-        std::vector<llama_token> doc_tokens = common_tokenize(vocab, document, false, true);
-
-        std::vector<llama_token> rerank_tokens = format_rerank(vocab, query_tokens, doc_tokens);
-
-        llama_memory_clear(llama_get_memory(ctx), false);
-
-        // Process the rerank input
-        try {
-            params.prompt = tokens_to_str(ctx, rerank_tokens.begin(), rerank_tokens.end());
-            initSampling();
-            loadPrompt({}); // No media paths for rerank
-            beginCompletion();
-            doCompletion();
-
-            // Get the rerank score (single embedding value for rank pooling)
-            float *data = llama_get_embeddings_seq(ctx, 0);
-            if (data) {
-                scores.push_back(data[0]); // For rank pooling, the score is the first (and only) dimension
-            } else {
-                scores.push_back(-1e6f); // Default low score if computation failed
-            }
-        } catch (const std::exception &e) {
-            LOG_WARNING("rerank computation failed for document %zu: %s", i, e.what());
-            scores.push_back(-1e6f);
-        }
-        endCompletion();
-
-        // Clear KV cache again to prepare for next document or restore original state
-        llama_memory_clear(llama_get_memory(ctx), false);
-    }
-
-    return scores;
-}
-
-std::string llama_rn_context::bench(int pp, int tg, int pl, int nr)
-{
-    if (is_predicting) {
-        LOG_ERROR("cannot benchmark while predicting", "");
-        return std::string("[]");
-    }
-
-    is_predicting = true;
-
-    double pp_avg = 0;
-    double tg_avg = 0;
-
-    double pp_std = 0;
-    double tg_std = 0;
-
-    // TODO: move batch into llama_rn_context (related https://github.com/mybigday/llama.rn/issues/30)
-    llama_batch batch = llama_batch_init(
-        std::min(pp, params.n_ubatch), // max n_tokens is limited by n_ubatch
-        0,                         // No embeddings
-        1                          // Single sequence
-    );
-
-    for (int i = 0; i < nr; i++)
-    {
-        llama_batch_clear(&batch);
-
-        const int n_tokens = pp;
-
-        for (int i = 0; i < n_tokens; i++)
-        {
-            llama_batch_add(&batch, 0, i, {0}, false);
-        }
-        batch.logits[batch.n_tokens - 1] = 1; // true
-
-        llama_memory_clear(llama_get_memory(ctx), true);
-
-        const int64_t t_pp_start = llama_time_us();
-        if (llama_decode(ctx, batch) != 0)
-        {
-            LOG_ERROR("llama_decode() failed during prompt", "");
-        }
-        const int64_t t_pp_end = llama_time_us();
-
-        llama_memory_clear(llama_get_memory(ctx), true);
-
-        if (is_interrupted) break;
-
-        const int64_t t_tg_start = llama_time_us();
-
-        for (int i = 0; i < tg; i++)
-        {
-            llama_batch_clear(&batch);
-
-            for (int j = 0; j < pl; j++)
-            {
-                llama_batch_add(&batch, 0, i, {j}, true);
-            }
-
-            if (llama_decode(ctx, batch) != 0)
-            {
-                LOG_ERROR("llama_decode() failed during text generation", "");
-            }
-            if (is_interrupted) break;
-        }
-
-        const int64_t t_tg_end = llama_time_us();
-
-        llama_memory_clear(llama_get_memory(ctx), true);
-
-        const double t_pp = (t_pp_end - t_pp_start) / 1000000.0;
-        const double t_tg = (t_tg_end - t_tg_start) / 1000000.0;
-
-        const double speed_pp = pp / t_pp;
-        const double speed_tg = (pl * tg) / t_tg;
-
-        pp_avg += speed_pp;
-        tg_avg += speed_tg;
-
-        pp_std += speed_pp * speed_pp;
-        tg_std += speed_tg * speed_tg;
-    }
-
-    pp_avg /= nr;
-    tg_avg /= nr;
-
-    if (nr > 1) {
-        pp_std = sqrt(pp_std / (nr - 1) - pp_avg * pp_avg * nr / (nr - 1));
-        tg_std = sqrt(tg_std / (nr - 1) - tg_avg * tg_avg * nr / (nr - 1));
-    } else {
-        pp_std = 0;
-        tg_std = 0;
-    }
-
-    if (is_interrupted) llama_memory_clear(llama_get_memory(ctx), true);
-    endCompletion();
-
-    char model_desc[128];
-    llama_model_desc(model, model_desc, sizeof(model_desc));
-    return std::string("[\"") + model_desc + std::string("\",") +
-        std::to_string(llama_model_size(model)) + std::string(",") +
-        std::to_string(llama_model_n_params(model)) + std::string(",") +
-        std::to_string(pp_avg) + std::string(",") +
-        std::to_string(pp_std) + std::string(",") +
-        std::to_string(tg_avg) + std::string(",") +
-        std::to_string(tg_std) +
-        std::string("]");
-}
-
 int llama_rn_context::applyLoraAdapters(std::vector<common_adapter_lora_info> lora) {
     for (auto &la : lora) {
         la.ptr = llama_adapter_lora_init(model, la.path.c_str());
@@ -878,56 +269,6 @@ bool llama_rn_context::initMultimodal(const std::string &mmproj_path, bool use_g
         LOG_ERROR("[DEBUG] Failed to initialize multimodal: %s", e.what());
         return false;
     }
-}
-
-void llama_rn_context::processMedia(
-    const std::string &prompt,
-    const std::vector<std::string> &media_paths
-) {
-    if (!isMultimodalEnabled()) {
-        throw std::runtime_error("Multimodal is not enabled but image paths are provided");
-    }
-
-    // Delegate to the mtmd_wrapper method
-    mtmd_wrapper->processMedia(
-        ctx,
-        prompt,
-        media_paths,
-        n_ctx,
-        params.n_batch,
-        n_past,
-        embd,
-        context_full,
-        ctx_sampling
-    );
-}
-
-llama_rn_tokenize_result llama_rn_context::tokenize(const std::string &text, const std::vector<std::string> &media_paths) {
-    if (media_paths.size() > 0) {
-        if (!isMultimodalEnabled()) {
-            throw std::runtime_error("Multimodal is not enabled but media paths are provided");
-        }
-        auto result = tokenizeWithMedia(mtmd_wrapper, text, media_paths);
-        mtmd_input_chunks_free(result.chunks);
-        llama_rn_tokenize_result tokenize_result = {
-            .tokens = result.tokens,
-            .has_media = true,
-            .bitmap_hashes = result.bitmap_hashes,
-            .chunk_pos = result.chunk_pos,
-            .chunk_pos_media = result.chunk_pos_media,
-        };
-        return tokenize_result;
-    }
-    std::vector<llama_token> text_tokens;
-    text_tokens = common_tokenize(ctx, text, false);
-    llama_rn_tokenize_result tokenize_result = {
-        .tokens = text_tokens,
-        .has_media = false,
-        .bitmap_hashes = {},
-        .chunk_pos = {},
-        .chunk_pos_media = {},
-    };
-    return tokenize_result;
 }
 
 bool llama_rn_context::isMultimodalEnabled() const {
