+++ cpp/rn-llama.h
@@ -29,44 +29,17 @@ std::string tokens_to_str(llama_context *ctx, const std::vector<llama_token>::co
 
 lm_ggml_type kv_cache_type_from_str(const std::string & s);
 
-enum stop_type
-{
-    STOP_FULL,
-    STOP_PARTIAL,
-};
-
-// completion token output with probabilities
-struct completion_token_output
-{
-    struct token_prob
-    {
-        llama_token tok;
-        float prob;
-    };
-
-    std::vector<token_prob> probs;
-    llama_token tok;
-};
-
-struct completion_partial_output
-{
-  std::string content;
-  std::string reasoning_content;
-  std::vector<common_chat_tool_call> tool_calls;
-  std::string accumulated_text;
-};
+// Forward declarations - actual definitions are in rn-completion.h
+// Note: enum forward declarations not allowed in C++, using include in implementation file
+struct completion_token_output;
+struct completion_partial_output;
+struct llama_rn_tokenize_result;
 
 struct llama_rn_context_mtmd;
 
 struct llama_rn_context_tts;
 
-struct llama_rn_tokenize_result {
-    std::vector<llama_token> tokens;
-    bool has_media = false;
-    std::vector<std::string> bitmap_hashes;
-    std::vector<size_t> chunk_pos; // both text and media
-    std::vector<size_t> chunk_pos_media; // media only
-};
+struct llama_rn_context_completion;
 
 // Main context class
 struct llama_rn_context {
@@ -79,34 +52,12 @@ struct llama_rn_context {
     llama_context *ctx = nullptr;
     common_chat_templates_ptr templates;
     int n_ctx;
-    common_sampler *ctx_sampling = nullptr;
 
-    // Completion state fields
-    bool is_predicting = false;
-    bool is_interrupted = false;
-    bool has_next_token = false;
-    std::string generated_text;
-    std::vector<completion_token_output> generated_token_probs;
-    size_t num_prompt_tokens = 0;
-    size_t num_tokens_predicted = 0;
-    llama_pos n_past = 0;
-    size_t n_remain = 0;
-    std::vector<llama_token> embd;
-    bool incomplete = false;
-    bool context_full = false;
-    bool truncated = false;
-    bool stopped_eos = false;
-    bool stopped_word = false;
-    bool stopped_limit = false;
-    std::string stopping_word;
-    // Current completion parameters for chat parsing
-    int current_chat_format = COMMON_CHAT_FORMAT_CONTENT_ONLY;
-    common_reasoning_format current_reasoning_format = COMMON_REASONING_FORMAT_NONE;
-    bool current_thinking_forced_open = false;
+    // Completion context
+    llama_rn_context_completion *completion = nullptr;
 
     ~llama_rn_context();
 
-    void rewind();
     bool loadModel(common_params &params_);
 
     // Model methods
@@ -127,26 +78,6 @@ struct llama_rn_context {
       const std::string &messages,
       const std::string &chat_template
     ) const;
-    llama_rn_tokenize_result tokenize(const std::string &text, const std::vector<std::string> &media_paths);
-
-    // Completion processing methods
-    bool initSampling();
-    void truncatePrompt(std::vector<llama_token> &prompt_tokens);
-    void loadPrompt(const std::vector<std::string> &media_paths);
-    void beginCompletion();
-    void beginCompletion(int chat_format, common_reasoning_format reasoning_format, bool thinking_forced_open);
-    void endCompletion();
-    completion_token_output nextToken();
-    size_t findStoppingStrings(const std::string &text, const size_t last_token_size, const stop_type type);
-    completion_token_output doCompletion();
-    completion_partial_output getPartialOutput(const std::string &token_text);
-
-    // Embedding methods
-    std::vector<float> getEmbedding(common_params &embd_params);
-    std::vector<float> rerank(const std::string &query, const std::vector<std::string> &documents);
-
-    // Benchmarking methods
-    std::string bench(int pp, int tg, int pl, int nr);
 
     // Lora methods
     std::vector<common_adapter_lora_info> lora;
@@ -162,12 +93,6 @@ struct llama_rn_context {
     bool isMultimodalSupportVision() const;
     bool isMultimodalSupportAudio() const;
     void releaseMultimodal();
-
-    // Multimodal processing methods
-    void processMedia(
-      const std::string &prompt,
-      const std::vector<std::string> &media_paths
-    );
 
     // TTS fields and methods (delegated to TTS context)
     llama_rn_context_tts *tts_wrapper = nullptr;
